+++ x/vesting/keeper/msg_server_test.go
@@ -11,6 +11,8 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	vestingexported "github.com/cosmos/cosmos-sdk/x/auth/vesting/exported"
 	sdkvesting "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
+	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/evmos/evmos/v20/testutil"
 	"github.com/evmos/evmos/v20/testutil/integration/evmos/network"
 	utiltx "github.com/evmos/evmos/v20/testutil/tx"
@@ -20,15 +22,16 @@ import (
 )
 
 var (
-	vestAmount     = int64(1000)
-	balances       = sdk.NewCoins(sdk.NewInt64Coin(utils.BaseDenom, vestAmount))
-	quarter        = sdk.NewCoins(sdk.NewInt64Coin(utils.BaseDenom, 250))
-	addr3          = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
-	addr4          = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
-	funder         = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
-	vestingAddr    = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
-	lockupPeriods  = sdkvesting.Periods{{Length: 5000, Amount: balances}}
-	vestingPeriods = sdkvesting.Periods{
+	vestAmount      = int64(1000)
+	balances        = sdk.NewCoins(sdk.NewInt64Coin(utils.BaseDenom, vestAmount))
+	delegationCoins = sdk.NewCoins(sdk.NewInt64Coin(utils.BaseDenom, 1e18))
+	quarter         = sdk.NewCoins(sdk.NewInt64Coin(utils.BaseDenom, 250))
+	addr3           = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
+	addr4           = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
+	funder          = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
+	vestingAddr     = sdk.AccAddress(utiltx.GenerateAddress().Bytes())
+	lockupPeriods   = sdkvesting.Periods{{Length: 5000, Amount: balances}}
+	vestingPeriods  = sdkvesting.Periods{
 		{Length: 2000, Amount: quarter},
 		{Length: 2000, Amount: quarter},
 		{Length: 2000, Amount: quarter},
@@ -49,8 +52,13 @@ func TestMsgFundVestingAccount(t *testing.T) {
 		// preFundClawback determines if the clawback vesting account should be already be funded before the test case
 		// this is used to test merging new vesting amounts to existing lockup and vesting schedules
 		preFundClawback bool
-		expPass         bool
-		errContains     string
+		// preExistentDelegation determines if the account should have a delegation before being converted to vesting account
+		preExistentDelegation bool
+		// withDelegation determines if the account should have a delegation after being converted to vesting account
+		withDelegation   bool
+		expPass          bool
+		errContains      string
+		expDelegatedFree sdk.Coins
 	}{
 		{
 			name:         "pass - lockup and vesting defined",
@@ -85,6 +93,7 @@ func TestMsgFundVestingAccount(t *testing.T) {
 			vesting:      vestingPeriods,
 			initClawback: false,
 			expPass:      false,
+			errContains:  "account is not subject to clawback vesting",
 		},
 		{
 			name:               "true - fund existing vesting account",
@@ -96,6 +105,42 @@ func TestMsgFundVestingAccount(t *testing.T) {
 			initClawback:       true,
 			preFundClawback:    true,
 			expPass:            true,
+		},
+		{
+			name:                  "true - create vesting account with pre existing delegation",
+			funder:                funder,
+			vestingAddr:           vestingAddr,
+			lockup:                lockupPeriods,
+			vesting:               vestingPeriods,
+			initClawback:          true,
+			preExistentDelegation: true,
+			expPass:               true,
+		},
+		{
+			name:                  "true - fund existing vesting account with pre existing delegation",
+			funder:                funder,
+			vestingAddr:           vestingAddr,
+			lockup:                lockupPeriods,
+			vesting:               vestingPeriods,
+			expectExtraBalance:    vestAmount,
+			initClawback:          true,
+			preExistentDelegation: true,
+			preFundClawback:       true,
+			expPass:               true,
+		},
+		{
+			name:                  "true - fund existing vesting account with delegations pre and post creation",
+			funder:                funder,
+			vestingAddr:           vestingAddr,
+			lockup:                lockupPeriods,
+			vesting:               vestingPeriods,
+			expectExtraBalance:    vestAmount,
+			initClawback:          true,
+			preExistentDelegation: true,
+			withDelegation:        true,
+			preFundClawback:       true,
+			expPass:               true,
+			expDelegatedFree:      delegationCoins,
 		},
 	}
 	for _, tc := range testCases {
@@ -109,6 +154,17 @@ func TestMsgFundVestingAccount(t *testing.T) {
 			require.NoError(t, err, "failed to fund target account")
 			err = nw.App.BankKeeper.SendCoins(ctx, tc.vestingAddr, tc.funder, balances)
 			require.NoError(t, err, "failed to send tokens to funder account")
+
+			if tc.preExistentDelegation {
+				// in order to delegate from the future vesting account, we need to
+				// send it some funds
+				err = testutil.FundAccount(ctx, nw.App.BankKeeper, tc.vestingAddr, delegationCoins)
+				require.NoError(t, err, "failed to fund vesting account")
+				msgDelegate := stakingtypes.NewMsgDelegate(tc.vestingAddr.String(), nw.GetValidators()[0].OperatorAddress, delegationCoins[0])
+				msgSrv := stakingkeeper.NewMsgServerImpl(nw.App.StakingKeeper.Keeper)
+				_, err = msgSrv.Delegate(ctx, msgDelegate)
+				require.NoError(t, err, "failed to delegate")
+			}
 
 			// create a clawback vesting account if necessary
 			if tc.initClawback {
@@ -130,6 +186,17 @@ func TestMsgFundVestingAccount(t *testing.T) {
 				require.NoError(t, err, "failed to fund vesting account")
 			}
 
+			if tc.withDelegation {
+				// in order to delegate from the vesting account, we need to
+				// send it some funds
+				err = testutil.FundAccount(ctx, nw.App.BankKeeper, tc.vestingAddr, delegationCoins)
+				require.NoError(t, err, "failed to fund vesting account")
+				msgDelegate := stakingtypes.NewMsgDelegate(tc.vestingAddr.String(), nw.GetValidators()[0].OperatorAddress, delegationCoins[0])
+				msgSrv := stakingkeeper.NewMsgServerImpl(nw.App.StakingKeeper.Keeper)
+				_, err = msgSrv.Delegate(ctx, msgDelegate)
+				require.NoError(t, err, "failed to delegate")
+			}
+
 			// fund the vesting account
 			msg := types.NewMsgFundVestingAccount(
 				tc.funder,
@@ -144,6 +211,7 @@ func TestMsgFundVestingAccount(t *testing.T) {
 			expRes := &types.MsgFundVestingAccountResponse{}
 			balanceFunder := nw.App.BankKeeper.GetBalance(ctx, tc.funder, utils.BaseDenom)
 			balanceVestingAddr := nw.App.BankKeeper.GetBalance(ctx, tc.vestingAddr, utils.BaseDenom)
+			spendableBalanceVestingAddr := nw.App.BankKeeper.SpendableCoin(ctx, tc.vestingAddr, utils.BaseDenom)
 
 			if tc.expPass {
 				require.NoError(t, err, tc.name)
@@ -151,9 +219,14 @@ func TestMsgFundVestingAccount(t *testing.T) {
 
 				accI := nw.App.AccountKeeper.GetAccount(ctx, tc.vestingAddr)
 				require.NotNil(t, accI)
-				require.IsType(t, &types.ClawbackVestingAccount{}, accI)
+				vestAcc, ok := accI.(*types.ClawbackVestingAccount)
+				require.True(t, ok)
+
 				require.Equal(t, sdk.NewInt64Coin(utils.BaseDenom, 0), balanceFunder)
 				require.Equal(t, sdk.NewInt64Coin(utils.BaseDenom, vestAmount+tc.expectExtraBalance), balanceVestingAddr)
+				require.Equal(t, tc.expDelegatedFree, vestAcc.DelegatedFree)
+				require.Empty(t, vestAcc.DelegatedVesting)
+				require.True(t, spendableBalanceVestingAddr.Amount.IsZero())
 			} else {
 				require.Error(t, err, tc.name)
 				require.ErrorContains(t, err, tc.errContains)
@@ -293,6 +366,25 @@ func TestMsgCreateClawbackVestingAccount(t *testing.T) {
 			vestingAddr: vestingAddr,
 			expPass:     true,
 		},
+		{
+			name: "success - with pre-existent delegation",
+			malleate: func(funder sdk.AccAddress, vestingAddr sdk.AccAddress) {
+				// fund the funder and vesting accounts from Bankkeeper
+				err := testutil.FundAccount(ctx, nw.App.BankKeeper, funder, balances)
+				require.NoError(t, err)
+				err = testutil.FundAccount(ctx, nw.App.BankKeeper, vestingAddr, delegationCoins)
+				require.NoError(t, err, "failed to fund vesting account")
+
+				// create a delegation for the vestingAcc
+				msgDelegate := stakingtypes.NewMsgDelegate(vestingAddr.String(), nw.GetValidators()[0].OperatorAddress, delegationCoins[0])
+				msgSrv := stakingkeeper.NewMsgServerImpl(nw.App.StakingKeeper.Keeper)
+				_, err = msgSrv.Delegate(ctx, msgDelegate)
+				require.NoError(t, err, "failed to delegate")
+			},
+			funder:      funderAddr,
+			vestingAddr: vestingAddr,
+			expPass:     true,
+		},
 	}
 
 	for _, tc := range testcases {
@@ -312,8 +404,11 @@ func TestMsgCreateClawbackVestingAccount(t *testing.T) {
 				require.Equal(t, &types.MsgCreateClawbackVestingAccountResponse{}, res)
 
 				accI := nw.App.AccountKeeper.GetAccount(ctx, tc.vestingAddr)
+				vestAcc, ok := accI.(*types.ClawbackVestingAccount)
 				require.NotNil(t, accI, "expected account to be created")
-				require.IsType(t, &types.ClawbackVestingAccount{}, accI, "expected account to be a clawback vesting account")
+				require.True(t, ok, "expected account to be a clawback vesting account")
+				require.Empty(t, vestAcc.DelegatedFree)
+				require.Empty(t, vestAcc.DelegatedVesting)
 			} else {
 				require.Error(t, err)
 				require.ErrorContains(t, err, tc.errContains)